Move group/role code into its own file
This is so we can access the group creation logic outside of a management command, and for general cleanliness.
