+++ user/models.py
@@ -7,12 +7,14 @@ from django.contrib.auth.password_validation import validate_password
 
 from rest_framework import serializers
 
+from user.groups import (
+    GROUPS,
+    EDITOR_GROUP,
+    COPYEDITOR_GROUP,
+    CONTRIBUTOR_GROUP,
+    role_at_least,
+)
 
-GROUPS = {
-    "Contributor": [],
-    "Copyeditor": ["Contributor"],
-    "Editor": ["Copyeditor", "Contributor"],
-}
 FORBIDDEN_USERNAMES = {"admin", "administrator", "root", "toor", "sudo", "sudoers"}
 
 
@@ -26,12 +28,12 @@ class SluglineUser(AbstractUser):
     @property
     def role(self):
         """Returns the highest role that a user has."""
-        if self.groups.filter(name="Editor").exists():
-            return "Editor"
-        elif self.groups.filter(name="Copyeditor").exists():
-            return "Copyeditor"
+        if self.groups.filter(name=EDITOR_GROUP).exists():
+            return EDITOR_GROUP
+        elif self.groups.filter(name=COPYEDITOR_GROUP).exists():
+            return COPYEDITOR_GROUP
         else:
-            return "Contributor"
+            return CONTRIBUTOR_GROUP
 
     # We write a setter as we construct temporary users when doing password validation, and sometimes role information
     # is part of the data.
@@ -41,7 +43,7 @@ class SluglineUser(AbstractUser):
 
     def at_least(self, role):
         """Returns if a user has at least a given role's privileges"""
-        return self.is_staff or role == self.role or role in GROUPS.get(self.role, [])
+        return self.is_staff or role_at_least(self.role, role)
 
     class Meta:
         ordering = ["date_joined"]
